Implemented a new version of the library. Now all the specialization vector and matrix classes are placed in 'Vec.hpp' and 'Mat.hpp' header files, respectively. This is done to avoid possible violations of the One Definition Rule while also preserving the general 'Vec' and 'Mat' template classes.

+++ dev/numa/src/Random.cpp
@@ -2,10 +2,9 @@
 
 #include <random>
 
-namespace numa
-{
-	double RandomDouble()
-	{
+namespace numa {
+
+	double RandomDouble() {
 		// Returns a random real in the [0,1) range.
 		// return rand() / (RAND_MAX + 1.0); // C-style alternative
 
@@ -13,46 +12,39 @@ namespace numa
 		static std::mt19937 generator{}; // static std::mt19937 generator;
 		return distribution(generator);
 	}
-	double RandomDouble(double min, double max)
-	{
+	double RandomDouble(double min, double max) {
 		// Returns a random real in [min,max).
 		return min + (max - min) * RandomDouble();
 	}
 
-	float RandomFloat()
-	{
+	float RandomFloat() {
 		return static_cast<float>(RandomDouble());
 	}
-	float RandomFloat(double min, double max)
-	{
+	float RandomFloat(double min, double max) {
 		return static_cast<float>(RandomDouble(min, max));
 	}
 
-	Vec2 RandomVec2()
-	{
+	Vec2 RandomVec2() {
 		return Vec2{
 			RandomDouble(),
 			RandomDouble()
 		};
 	}
-	Vec2 RandomVec2(double min, double max)
-	{
+	Vec2 RandomVec2(double min, double max) {
 		return Vec2{
 			RandomDouble(min, max),
 			RandomDouble(min, max)
 		};
 	}
 
-	Vec3 RandomVec3()
-	{
+	Vec3 RandomVec3() {
 		return Vec3{
 			RandomDouble(),
 			RandomDouble(),
 			RandomDouble()
 		};
 	}
-	Vec3 RandomVec3(double min, double max)
-	{
+	Vec3 RandomVec3(double min, double max) {
 		return Vec3{
 			RandomDouble(min, max),
 			RandomDouble(min, max),
@@ -60,39 +52,32 @@ namespace numa
 		};
 	}
 
-	Vec2 RandomInUnitSquare()
-	{
+	Vec2 RandomInUnitSquare() {
 		return RandomVec2(-1.0, 1.0);
 	}
-	Vec2 RandomInSquare(double side)
-	{
-		return RandomInUnitSquare() * side;
+	Vec2 RandomInSquare(double side) {
+		return RandomInUnitSquare() * static_cast<float>(side); // TEST cast
 	}
 
-	Vec3 RandomInUnitCube()
-	{
+	Vec3 RandomInUnitCube() {
 		return RandomVec3(-1.0, 1.0f);
 	}
-	Vec3 RandomInCube(double side)
-	{
-		return RandomInUnitCube() * side;
+	Vec3 RandomInCube(double side) {
+		return RandomInUnitCube() * static_cast<float>(side); // TEST cast
 	}
 
-	Vec3 RandomInUnitSphere()
-	{
-		while (true)
-		{
+	Vec3 RandomInUnitSphere() {
+		while (true) {
 			Vec3 randomVec3 = RandomInUnitCube();
 			if (Length2(randomVec3) <= 1.0f)
 				return randomVec3;
 		}
 	}
-	Vec3 RandomOnUnitSphere()
-	{
+	Vec3 RandomOnUnitSphere() {
 		return Normalize(RandomInUnitCube());
 	}
-	Vec3 RandomInSphere(double radius)
-	{
-		return RandomInUnitSphere() * radius;
+	Vec3 RandomInSphere(double radius) {
+		return RandomInUnitSphere() * static_cast<float>(radius); // TEST cast
 	}
+
 }
